requirements.txt
mido>=1.3.2
Pillow>=11.0.0
imageio>=2.37.0
imageio-ffmpeg>=0.5.1
numpy>=2.0.0
